Better Implement Main View
The main view is now broken into the menus, label, and button. And then the List and Source fxml views are loaded in. 
SourceView got renamed to AddSource view to properly reflect its purpose.
The menus/buttons and labels were given alignments within the cell in order to keep them from going all over the place and looking ugly.